Use bang-methods as actions for commands
